+++ src/App.tsx
@@ -1,9 +1,9 @@
 import React, { useState } from 'react';
 import './app.scss';
 import { createBrowserRouter, RouterProvider } from 'react-router';
-import HomePage from '../pages/home-page/HomePage';
-import NotFoundPage from '../pages/NotFoundPage';
-import DetailedPage from '../pages/detailed-page/DetailedPage';
+import HomePage from './pages/home-page/HomePage';
+import NotFoundPage from './pages/NotFoundPage';
+import CardDetailsContainer from './components/card-details-container/card-details-container';
 
 const App: React.FC = () => {
   const [currentPage] = useState(1);
@@ -13,7 +13,7 @@ const App: React.FC = () => {
       path: '/',
       element: <HomePage currentPage={currentPage} />,
       errorElement: <NotFoundPage />,
-      children: [{ path: 'details/:id/', element: <DetailedPage /> }],
+      children: [{ path: 'details/:id/', element: <CardDetailsContainer /> }],
     },
     { path: '*', element: <NotFoundPage /> },
   ]);
