+++ src/components/Core/core.scss
