+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import { StrictMode } from 'react';
 import './index.scss';
 import ErrorBoundary from './components/error-boundary/Error-boundary';
-import App from './components/App';
+import App from './App';
 
 const rootElement = document.getElementById('root');
